WIP of ECDSA signatures working.
This at least proves that ECDSA signatures are doable. This does not re-encode signatures correctly in all cases when concating the octets. The concatenation depends on the curve used to know whether or not padding is required, and actually to trim off sign bytes. But, this at least gets the test passing.