Simplified parsing.
This addresses a number of complexity issues.

First, we greatly decreased the number of "segements" we were making. We were going from A, to B, and back to A again.

Second, we remove the handling of content data from the decoder. The decoder itself isn't fully smart enough to handle content when unknown-length contents were added, and left it up to the element parser. This fully shifts the responsibility to the parser.